Decoupled styles from components
