+++ .env
@@ -1,0 +1,1 @@
+NODE_PATH='./src/'
