+++ src/Pages/Rewards/index.js
